pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "localpilot"
version = "1.0.0"
description = "LocalPilot — offline LLM + RAG desktop application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "llama-cpp-python==0.3.2",
    "scikit-learn>=1.5.0",
    "joblib>=1.4.0",
    "pymupdf>=1.24.0",
    "Pillow>=10.3.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "pyinstaller>=6.6.0",
    "openai>=2.24.0",
    "customtkinter>=5.2.2",
]

[project.optional-dependencies]
ocr = ["pytesseract>=0.3.10"]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
localpilot = "localpilot.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/localpilot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
